+++ Changelog.md
@@ -22,3 +22,14 @@
 + intro
   1. modified the floor indicators(lcds) to display the 1st and 60th floor correctly
   2. ​
+
+## 3. Version 0.3
+
+*Date*: 2017-7-21
+
++ Intro
+    1. add some features in object ele:
+        + add des list storing the destinations of passengers
+        + add function to get the current amount of passengers inside the ele
+
+    2. add commands_function to the schedule.py